got rid of unit types DM Walking Corpse and DM Soulless
I have not removed them yet as it would break backwards compatibilty.
For now they are now based on core units but still advance to Running
Corpse. Zombies created in scenarios are now standard core units.
Now I should write temporary code to transform existing custom zobmies
to the standard ones.
As for why I have done this, it was not necessary to create custom units
   in the first place. They were just duplicates of core units (with the
exception of additional units in line but that can be done with
[advancefrom]).